PicaEncoder implemented
The PicaEncoder encodes the "Record-Field-Subfields" Stream into pica+
Format. If there are S-subfields in the input stream which is processed
by PicaDecoder (by default) it will be not able to encode them to
original format (because of lost of data during the decoding). In this
case, the decoder should use the option appendControlSubField="false" to
keep the original data.