Fixes #125: Set correct file modes in distribution on Linux
Ensures that directory and file permissions are set correctly in the
distribution package. Additionally, the line endings of the shell and
batch scripts are converted to the one of the target platform. 

This commit adds tar.gz as an additional distribution file format since
directory modes do not work correctly in zip-files created by the
assembly plugin (see http://jira.codehaus.org/browse/MASSEMBLY-494 for
an explanation and a workaround if desired). However, tar.gz files can
be cumbersome to work with on Windows. Consequently, both archive
formats should be offered.