Fix #137: Refactored the pica parser to be more lenient.
The parser does expect fields and subfields to end with a field or
subfield marker anymore. This should allow parsing of pica serialisation
variants which use the markers as delimiters as well as variants which
use the markers as indicators.

The fixUnexpectedEOR option was removed as there is no unexpected end of
record anymore.