Handles record markers in pica records gracefully (fixes #139).
Modified the pica parser to behave as if a field marker was encountered
when a record marker is encountered. Strictly speaking this is not
correct behaviour but the PicaDecoder expects single records in the
first place. Consequently a record marker should only be found at the
beginning or the end of the input.