New Collector EqualsFilter
With the EqualsFilter-Collector a literal is only emitted if all
received values are equal.

The following example collects values which are stored in separate
literals (037H.d and 037H.e). Then it checks if both values are equal.
If the values are equal the value is emitted. If the values are not
equal no value is emitted.

<equalsFilter name="equalsFilterd" value="${one}">
	<data source="037H.d" name="one" />
	<data source="037H.e" name="two" />
</equalsFilter>