Added support for conditional activation of collectors.
This commits introduces an <if> statement which can be used to control
under which conditions a collector generates output. It can be used for
example to write a combine statement that only generates outputif data3
has a value of "x":

<combine name="output" value="${a}${b}">
  <if>
    <data source="data3">
      <equals string="x" />
    </data>
  </if>
  <data source="data1" name="a" />
  <data source="data2" name="b" />
</combine>

The quantors introduced by commits 475a55fc94, 12a0537220 and
e65f6d2e1a are meant to be used inside the <if> statement.