Added header, footer and separator settings to ObjectWriter.
These settings allow to prefix the output with a header string and to
append a footer at the end of the stream. Additionally, the separator
between objects can be now changed. Previously, this was always a
newline character.

The default settings of header, footer and separator are chosen so that
the output of ObjectWriter is the same as in the old version of the
module (i.e. a new line character is append to each object).

In addition to these new options the encoding and compression options
can now be set via Flux.

This commit offers a solution for issue #152: The output of the JSON
encoder can be turned into a JSON array by setting header to "[",
separator to ", " and footer to "]".