Added an <all> statement to metamorph.
The statement implements a universal quantor or conjunction operation
(logical and-operation). It allows for checking whether all statements
contained in the <all> statement generated an output. Using the <all>
statement conjunctions can be written in a more expressive way than
(ab)using the <combine> statement allows.

By default the <all> statement outputs an unamed literal with "true" as
value. However, this can be customized.

<all name="MyCustomName" value="all-fired">
	<data source="data1" />
	<data source="data2" />
</all>

The <all> statement supports reset and flushWith.

Conflicts:
	src/test/java/org/culturegraph/mf/morph/collectors/CollectorTest.java