Added an <any> statement to metamorph.
The statement implements a existential quantor or disjunction operation
(logical or-operation). It allows for checking whether at least one of
the statements contained in the <any> statement generated an output.
Using the <any> statement disjunctions can be written in a more
expressive way than (ab)using the <combine> statement allows.

By default the <any> statement outputs an unamed literal with "true" as
value. However, this can be customized.

<any name="MyCustomName" value="one-fired">
	<data source="data1" />
	<data source="data2" />
</any>

The <any> statement supports reset and flushWith.

Conflicts:
	src/test/java/org/culturegraph/mf/morph/collectors/CollectorTest.java