Added a <none> statement to metamorph.
The statement implements a logical not-operation. It allows for
checking whether none of the statements contained in the <none>
statement generated an output. Using the <none> statement negations can
be written in a more expressive way than (ab)using the <combine> and
<choose> statements allows.

By default the <none> statement outputs an unamed literal with "true" as
value. However, this can be customized.

<none name="MyCustomName" value="none-fired">
	<data source="data1" />
	<data source="data2" />
</none>

The <none> statement supports reset and flushWith.

Conflicts:
	src/test/java/org/culturegraph/mf/morph/collectors/CollectorTest.java