Refactored the dynamic module loading mechanism
The dynamic module loader no uses its own class loader which can add
files from multiple directories. It is based on the standard
URLClassLoader. 

The folder from which modules are loaded dynamically is no longer
provided via command line arguments but via a system property. The name
of the property is flux.pluginsdir; it expects an absolute path name.
The default setting for the module directory has been removed. If the
property is not set, no module directory is used.

In addition, a second directory whose contents are loaded dynamically,
is introduced for loading libraries typically provided by the user such
as JDBC drivers or JNDI providers. This directory is configured via the
flux.provideddir property.

The maven config has been changed to place libraries with scope provided
in the provided-libs dir instead of the standard libs dir. These
libraries are not added to the class path.

As first libary, the scope of the slf4j-log4j dependency has been
changed into a provided dependency.