Removed logging code from DirectoryClassLoader
The class loader is instantiated before dynamically loaded jars are
added. This means, if the slf4j backend implementation is loaded using
this mechanism (which should be the default), the backend is not yet
available when the DirectoryClassLoader is created. This will cause
slf4j to use the no operation logger which is not what the user expects.