Updates dependencies to latest version.
All dependencies are updated except for jdom 1.1 and antlr 3.4. Updating
the former to jdom2 would require all users of the JDomDocumentToStream
module to migrate to jdom2, too. We do not want to force users to do this.
So, we stick to jdom 1.

We still use antlr 3.4 on our build system to generate the Flux parser and
lexer. As there is no need to update this at the moment, we stay on
antlr 3.4.