Generate Flux parser and lexer as part of the build cycle
This commit updates the maven configuration to generate the code files for
the parser and lexer for Flux during the build process. The pre-generated
code files for the Flux parser and lexer are removed from the repository.

This change can cause peculiar errors if you do a `mvn clean` followed by
a `mvn install` while the project is opened in eclipse (or possibly any
other IDE that automatically compiles sources in the background). The
reason for this is that eclipe starts compiling the source files again as
soon as `mvn clean` completes. However, at this point the files generated
by antlr are not yet re-generated (and eclipse does not generate them) so
any files relying on these generated files will fail to compile.

The best approach to avoid problems during `mvn clean` is to close the
project in eclipse, run `mvn clean test` on the command line and only open
the project again once this has finished. Subsequently, eclipse background
complilation should work well.

This commit also updates the antlr version to 3.5.1.