Dynamic loading of slf4j binding does not work
See https://github.com/culturegraph/metafacture-runner/issues/2

This commit adds the log4j binding for slf4j to the class path.

Our attempts to load the slf4j binding dynamically from the
`provided-libs` dir does not work as expected. slf4j does not consider
the context class loader for loading its binding but relies on the class
loader which was used to load the classes of the slf4j api. Until a
solution is found for this problem, the binding need to be placed on the
class pathslf4j does not consider
the context class loader for loading its binding but relies on the class
loader which was used to load the classes of the slf4j api. Until a
solution is found for this problem, the binding need to be placed on the
class path.