Added source file location annotations to Metamorph DOM.
When debugging Metamorph scripts it is helpful to be able to link the
data, collector and function objects with their definition in the
Metamorph script. This commit annotates the XML DOM that is used to
build the actual metamorph object structure with location information
about location of each node in the source XML file.

The location information is attached as user data to element nodes using
the key `http://culturegraph.org/mf/user_data/location` (available via
`Location.USER_DATA_ID`). It contains start and end positions of the
element in the source file.

Implementing this functionality required a complete rewrite of the
`DOMLoader` because location information is only available from the SAX
parser and not from the DOM builder. Therefore the, implementation
uses a `Transformer` to construct a DOM from SAX events. The idea for
this implementation comes from
http://javacoalface.blogspot.de/2011/04/line-and-column-numbers-in-xml-dom.html

In the process of reimplementing the `DOMLoader` it was decided to move
the class and its helper classes out of the morph package and into its
own utility package due to the number of classes. Since we now have a
`util.xml` package the old `XMLUtil` class was moved into this package
as well.

Note: The default SAX implementation used by OpenJDK 1.7.0.60 includes
an old version of Xerces-J (2.7.1) which does not report location
information in x-included files correctly. A bug report for updating the
implementation exists (see https://bugs.openjdk.java.net/browse/JDK-8038043).
However, until this makes it into OpenJDK a newer version of Xerces needs
to be added manually to the classpath.