Refactored code for loading morph scripts.
In order to be able to include source file names in the location
information (see commit e10de395592d082596354646e555083e4aa5e1e6) the
XML input source needs to have the system id set to the file name of the
source file. This happens only if the `InputSource` object is
constructed from a URL. However, the old code for loading morph scripts
used `Reader`s or `InputStream`s for constructing the `InputSource`.

This commits rewrites the loading mechanism so that it works with URLs
instead of streams. Additionally, the number of `walk` methods in
`AbstractMetamorphDomWalker` are significantly reduced by moving the
creation of the `InputSource` to the `Metamorph` class.