Added source location infos to the morph pipeline.
Commit e10de395592d082596354646e555083e4aa5e1e6 annotated the Metamorph
DOM with information about the location of each element in the source
file. This commit adds this information to the data, collector and
function objects which are created based on the morph pipeline described
by the DOM.

The location information is made available through a new interface
called `KnowsSourceLocation`. It offers methods for setting and getting
location information from an object. `NamedValueReceiver` and
`NamedValueSource` extend this interface. The location information is
set via a setter method instead of via the constructor to avoid having
to add a constructor to every concrete collector or function class.