Formeta: Only accept allowed escape sequences
The Formeta parser used to allow any character in an escape sequence.
Most characters were simply replaced with themselves. This behaviour
makes the introduction of new escape codes in the future very difficult.
Thus, this commit changes the parser behaviour so that only the
currently used escape sequences (nr",:{}\ and whitespace) are allowed.