Simplified output of `PicaMultiscriptRemodeler`
The original version of `PicaMultiscriptRemodeler` used the original
script names for the intermediate layer of entities. This makes
processing of the output unnecessary complex because of the large number
of possible entity names. As currently only three types of scripts need
to be distinguished, the module was changed to map the script names to
one of the three script types which are then used as entity names.