Adds reader for multiline formeta records
The `FormetaDecoder` processes only individual records. If the
`LineReader` or `RecordReader` modules are used for splitting an input
stream consisting of a sequence of formeta records, this is quite
inconvenient as records either have to by in a single line or must be
seperated with a record separater character. As such a character is not
defined for the formeta format this is not a good solution.

However, the structure of formeta allows to recognise the end of a
record quite easily without having to properly parse the record. The new
module `FormetaRecordsReader` implements such a splitter. It splits a
sequence of formeta records between each record.