Bug fix: Allow different result types.
In commit 51f5c95bee3754651e30dd9f3e51c821e80221fe the `TestCase`
implementation was udpated to support any format for which a reader
exists to be used in the result section of a test case. However, the
commit did not include the necessary updates of the metamorph-test
schmema. These are added by this commit.