Add decoder for ISO 2709:2008 and MARC 21
The ISO 2709:2008 standard defines a framework for character-based
formats for exchanging (bibliographic) records. One instance of the
ISO standard is the MARC 21 standard for bibliographic records.

This commit adds a generic decoder for ISO 2709 records and a
Metafacture module for processing MARC 21 records which builds on the
generic decoder.

The new `Marc21Decoder' module will replace the old `MarcDecoder` in
the next major release of Metafacture. Until then the new decoder is
only available as in Java code but not in Flux.