Exclude check for boolean inversion from PMD
The argument for using bitwise inversion instead of boolean inversion is
quite dated and in tests I could not reproduce the improved performance
of bitwise inversion compared to boolean inversion. In any case, the
performance gain is so small in most cases that it does not compensate
for the loss in readability of the code.