Fix #204: High memory usage of Metamorph tests
The metamorph test framework loads all test definitions into memory
prior to running the tests. For each test definition the Metamorph
script under test is instantiated. If there are many test cases this can
consume a lot of memory (in particular if the script is large).

This is changed so that the Metamorph module is created in the `run`
function of `TestCase` and stored in a local variable.