Add FluxCommand annotation
The annotation marks classes as flux commands. The value of the
annotation specifies the name of the flux command.  The `FluxCommand`
annotation is a key feature of a new mechanism for declaring and loading
flux commands which is scheduled for integration into the next major
release of Metafacture. By introducing the annotation already in an
earlier release, module and plugin developers get a chance to prepare
their modules in time for the next major release.