Fix #219: Resolve x-includes in tests correctly
In the Metamorph test framework Metamorph transformations in external
files were not loaded by passing a URL to Metamorph but by passing a
Java IO reader. This prevented the XML parser in Metamorph from
resolving relative paths in x-includes correctly. This commit fixes this
problem by passing the URL of the external Metamorph transformation to
Metamorph.

Additionally, embedded Metamorph transformations did not have a valid
base URI which is needed for resolving x-includes with relatives paths
correctly.  This is fixed by setting the base URI of the embedded
transformation to the URI of the surrounding Metamorph-test document.

Without these changes relative x-includes were resolved based on the
current working directory which could yield to unexpected results and
does not work with Java resources in jars.

NOTE: This fix changes the way relative paths in x-includes are
resolved. Metamorph scripts containing x-includes with relative paths
may need to be updated. While previously the path had to be specified
relative to the current working directory it now needs to be specified
relative to the location of the including Metamorph script.
X-includes with absolute paths are not affected by this change.