Fix #4: Choose correct sed option for extended re
The GNU and BSD versions of sed use different options for enabling
extended regular expressions. In order for scripts to work on both
systems, the flux.sh script needs to check which sed version is
installed. This commits adds a check for this. Additionally, the
the script is aborted if sed does not seem to support extended
regexps at all.