Resolve #49, #210, #250: reset and sameEntity in choose
This commit enables the `reset` and `sameEntity` flags for the `choose`
collector (resolves #210 and resolves #250). In order for the `reset`
flag to work, the `choose` collector is changed to no longer always
clear its value after emitting it (fixes #49). To minimize the impact of
this modification on existing metamorph scripts, the `reset` flag is set
to true by default. Despite this the behaviour of `choose` is different
if it contains an `if`-condition. In the old implementation the chosen
value was only cleared after it was actually emitted (so only if the
condition was met). Now it is also cleared if it was attempted to be
emitted but was not due to the condition not being met. This is in-line
with the behaviour of the other collectors.