Disable javadoc doclint
The javadoc version in jdk 8 is much more strict about errors in javadoc
comments than earlier versions were. As a result the maven javadoc
plugin fails to create the javadoc jar when building a Metafacture
release.

This commit disables the strict checks in javadoc, so that we can
continue building Metafacture releases without having to fix all
comments first. This is only a workaround which should be removed once
the comments have been improved.