Fix usage of PrintStream
The test cases for `ObjectStdoutWriter` did not correctly handle the
redirected `System.out` stream.

When a `PrintStream` instance wraps an underlying
`ByteArrayOutputStream` instance, it is recommended to flush or close
the `PrintStream` before invoking the underlying instances's
`toString()`. Also, it is a good practice to call flush/close
explicitly.  This commit adds a flush method before calling
`toString()`.