Add module to turn records into entities
The `RecordToEntity` module replaces start-record and end-record events
with start- and end-entity events. This functionality is similar to the
behaviour of the `RecordBoundaryRemover` when created with an entity
name. However, the new module is more flexible as the entity name is
specified using a setter and the output of the id literal is also
configurable. Therefore, the single argument constructor in
`RecordBoundaryRemover` is marked as deprecated with the intention of
removing it in a future release.