Make whitespace removal in field names optional
Adds a configuration option to `PicaDecoder` to control whether leading
and trailing whitespace in field names is removed. By default it is
removed.

This commit changes the behaviour of the `PicaDecoder`. While older
versions remove all space characters in the field name (at any
position) the new version remove all whitespace but only at the start
or end of the field name.

Since pica field names do not normally contain whitespace, this change
is deemed to be insignificant.