Refactor error reporting in StreamValidator
Instead of throwing a `ValidationException` the module now invokes a
user-supplied error handler.

This makes it possible to directly throw an `AssertionError` when using
the `StreamValidator` in the Metamorph test runner. This solves the
problem of the old implementation which caught all subclasses of
`FormatException`. This resulted in difficult to understand error
messages when an error in the input data was reported as a failed
validation instead of broken data (fix #6).