Refactor Metamorph test framework
The `TestSuite` class is renamed to `MetamorphTestSuite` to make clear
that it is not part of core JUnit but a Metamorph specific test suite.
It also no longer extends JUnit's `Suite` but `ParentRunner` instead.

The interface of the test package is simplified. The only remaining
public class is `MetamorphTestSuite`. The visibility of all other
classes is reduced to package-private access. This allows to refactor
the implementation of the Metamorph test framework in coming releases.

The `TestConfigurationException` is removed and replaced with JUnit's
`InitializationError`.