Fix test names for Metamorph Test in Intellij
Intellij did not show the name of the xml files containing the tests but
only the string "xml". This was caused by Intellij interpreting the test
class name as a fully qualified java class name and attempting to
extract the class name from it. By making sure that the Metamorph Test
names do not contain any dots this problem can be avoided.