Refactor StreamBatchMerger and remove WrappingStreamPipe
The `WrappingStreamPipe` was used by `AbstractBatcher` only. It enabled
derived classes to intercept the events forwarded by the batcher module.
This functionality was only used by `StreamBatchMerger` and resulted in
a very complicated implementation.

By refactoring `StreamBatchMerger` to not derive from `AbstractBatcher`
its implementation is heavily simplified and it is no longer necessary
that `AbstractBatcher` implements `WrappingStreamPipe`.

This commit changes the interface of the `StreamBatchMerger` module. It
no longer returns the record and batch counts.