Refactor IdChangePipe and improve its Javadoc
Renames the parameters of the module to make them easier to understand.
The constructor which allowed to set the id literal is removed. The
setter method can be used instead. Javadocs for the class and all
parameters are added.  Additionally, the implementation of the module is
simplified by using `EntityPathTracker` instead of `StreamFlattener`.