Rename LiteralExctractor to LiteralToObject
The old name did not explain well what the module was doing. The new
name also better reflects that the module performs the reverse operation
of the  `ObjectToLiteral` module.

The Javadocs of the module are improved. The constructorsa re removed as
the pattern can also be set through a setter method.