Remove EventListSource
The design of the module does not fit well into the pipes-and-filters
architecture. It should have been implemented as converter which
receives a list of persisted stream events as input and emits these
events.

The module can be replaced by `StreamBuffer` in cases where it was used
together with `EventList` to buffer events.