Remove MultiOpener
The module does not work as there are no openers registered with the
`OpenerFactory` used by `MultiOpener` (and apparently never have been).

`MultiOpener` is removed instead of being fixed because the way the
concrete opener is chosen by the module  is not optimal. A future
implementation should work on URLs instead of requiring to explicitely
choose an opener implementation.