Remove modules in the reader package
These modules (expect for the `BeaconReader`) are merely combinations of
other modules and add no new functionality. They are often a source of
confusion as it is not clear for users whether they should use the decoders
or the readers. Additionally, the class hierarchy of the readers makes it
difficult to place them into independent packages as it is planned.

It is, however, not possible to remove all of the modules as the Metamorph
test framework depends on them for reading the test inputs and outputs.
Therefore, the readers used by the test framework are made part of the
`test` package.

The `BeaconReader` differs from the other readers as it is an actual
implementation of a beacon file reader. It is moved into the `converter`
package.

Users who use any of the other readers should replace them by the
corresponding combination of a record splitting module and a record
decoding module.

There is no replacement for the `MultiFormatReader`.