Framework: Remove dependency on stream package
The JavaDoc of `StreamReceiver` refered to the `WellformednessChecker`.
This introduced a dependency on the general streams package. As the
`WellformednessChecker` will be moved into the test tools and thus
effectively becomes a private module of the test framework, the
reference in the JavaDoc is removed.

Additionally, the JavaDocs of `StreamReceiver` are improved to better
reflect the behaviour implemented in existing nmodules.