Move helper interfaces used by XmlFilenameWriter
The `RecordIdentifier` and `FilenameExtractor` are not specific to xml
processing. Hence, they should not be in the `util.xml` package. As they
are currently only used by `XmlFilenameWriter` and as it is not clear
whether there will be any other module using them, the interfaces are
placed together with the `XmlFilenameWriter` in the `sinks` package.